Delegate all state management to models. #36
