+++ .nycrc
@@ -1,6 +1,6 @@
 {
   "include": [
-    "_src/*"
+    "src/*"
   ],
   "exclude": [
     "**/*test*"
