--- _src/index.js
@@ -1,4 +1,0 @@
-export { View as Form } from './form'
-export { View as Field } from './field'
-export { View as FieldSet } from './field-set'
-export { View as FieldArray } from './field-array'
