+++ src/field-set.js
@@ -1,37 +1,206 @@
-import { createElement } from 'react'
-import { oneOfType, func, object, string } from 'prop-types'
+import PropTypes from 'prop-types'
 import * as _ from './util'
 import * as SuperControl from './super-control'
 
-export class FieldSet extends SuperControl.View {
+export const Model = class FieldSetModel extends SuperControl.Model {
   constructor(...args) {
     super(...args)
-    this.ownProps = _.omit(this.props, [
-      'init', 'notify', 'validate', 'component'
-    ])
+    this.fields = {}
+    this._state.active = null
+    this.reset = this.reset.bind(this)
+    this.touch = this.touch.bind(this)
+    this.change = this.change.bind(this)
+    this.untouch = this.untouch.bind(this)
+    this.touchAll = this.touchAll.bind(this)
+    this.broadcast = this.broadcast.bind(this)
+    this.untouchAll = this.untouchAll.bind(this)
+    this.validateAll = this.validateAll.bind(this)
   }
-  getInit() {
-    return this.props.init
+  get values() {
+    return this.value
   }
-  modelField(...args) {
-    return modelFieldSet(...args)
+  get anyVisited() {
+    return this.isVisited
   }
-  render() {
-    const { props: { name, component }, ownProps: props } = this
-    if (_.isString(component)) {
-      return createElement(component, { ...props, name })
+  get anyTouched() {
+    return this.isTouched
+  }
+  get active() {
+    return this._state.active
+  }
+  get errors() {
+    const { fields, error } = this
+    return _.keys(fields)
+      .reduce((errors, key) => _.assign(
+        errors,
+        { [key]: fields[key].errors || fields[key].error }
+      ), { $self: error })
+  }
+  getState() {
+    return _.assign(super.getState(), _.pick(this, [
+      'values', 'anyVisited', 'anyTouched', 'active'
+    ]))
+  }
+  register(names, field) {
+    const [ first, ...rest ] = names
+    this.fields = rest.length
+      ? _.set(this.fields, [first], this.fields[first].register(rest, field))
+      : _.set(this.fields, [first], field)
+    return this._patch(names, field._state)
+  }
+  getField([ first, ...names ]) {
+    const child = this.fields[first]
+    if (!child) return null
+    if (names.length) {
+      return child instanceof FieldSetModel
+        ? child.getField(names)
+        : null
     }
-    return createElement(component, {
-      ...props,
-      name,
-      fields: this.model.toProp()
+    return child
+  }
+  unregister(names, field) {
+    const registered = this.getField(names)
+    if (registered !== field) return this
+    super._patch({
+      visits: -registered._state.visits,
+      touches: -registered._state.touches,
+      value: _.unset(this._state.value, names)
     })
+    const [ first, ...rest ] = names
+    this.fields = rest.length
+      ? _.set(this.fields, [first], this.fields[first].unregister(rest, field))
+      : _.unset(this.fields, [first])
+    return this
+  }
+  _patch(names, change, options = {}) {
+    if (!names.length) return super._patch(change, options)
+    const current = this._state
+    const next = _.omit(change, ['init', 'value'])
+    if ('init' in change && !this.form.isInitialized) {
+      next.init = _.set(current.init, names, change.init)
+    }
+    if ('value' in change) {
+      next.value = _.set(current.value, names, change.value)
+    }
+    if ('visits' in change && change.visits > 0) {
+      next.active = this.getField(names)
+    }
+    if ('touches' in change && change.touches > 0) {
+      next.active = null
+    }
+    super._patch(next, { ...options, silent: !!options.quiet })
+    const [ first, ...rest ] = names
+    const field = this.fields[first]
+    field instanceof FieldSetModel
+      ? field._patch(rest, change, options)
+      : field._patch(change, options)
+    return this
+  }
+  initialize(init) {
+    _.keys(init).forEach(key => {
+      this.fields[key] &&
+      this.fields[key].initialize(init[key])
+    })
+    super.initialize(init)
+  }
+  broadcast() {
+    this.publish()
+    this.eachField(field => _.invoke(field.broadcast || field.publish))
+    return this
+  }
+  change(path, value) {
+    const field = this.getField(_.toNames(path))
+    field && field.change(value)
+  }
+  touch(path) {
+    const field = this.getField(_.toNames(path))
+    field && field.touch()
+  }
+  untouch(path) {
+    const field = this.getField(_.toNames(path))
+    field && field.untouch()
+  }
+  touchAll() {
+    this.eachField(field => _.invoke(field.touchAll || field.touch))
+  }
+  untouchAll() {
+    this.eachField(field => _.invoke(field.untouchAll || field.untouch))
+  }
+  validate(options) {
+    const { form, config, value } = this
+    super._patch({ error: config.validate(value, form.values, this) }, options)
+  }
+  validateAll() {
+    this.validate()
+    this.eachField(field => _.invoke(field.validateAll || field.validate))
+    return this
+  }
+  reset(options) {
+    const { form, names, fields, init } = this
+    _.keys(fields).reverse().forEach(key => {
+      _.exists(init, key)
+        ? fields[key].reset(options)
+        : form.unregister(fields[key].names, fields[key])
+    })
+    form._patch(names, { value: init, error: null, notice: null }, options)
+  }
+  eachField(procedure) {
+    _.keys(this.fields).forEach(key => procedure(this.fields[key]))
+  }
+  static create(form, init = {}, route, config) {
+    return super.create(form, init, route, config)
+  }
+}
+
+export const View = class FieldSetView extends SuperControl.View {
+  constructor(...args) {
+    super(...args)
+    this.register = this.register.bind(this)
+  }
+  get Model() {
+    return Model
+  }
+  register({ route, ...params }) {
+    return this.context['@@super-controls'].register({
+      ...params,
+      route: [_ => this.props.name, ...route]
+    })
+  }
+  getChildContext() {
+    return { '@@super-controls': { register: this.register } }
+  }
+  equalState(current, next) {
+    const { init: stateInit, values: stateValues, ...restState } = current
+    const { init: nextInit, values: nextValues, ...restNext } = next
+    return _.shallowEqual(restState, restNext) &&
+           _.shallowEqual(stateValues, nextValues) &&
+           _.shallowEqual(stateInit, nextInit)
+  }
+  componentWillReceiveProps(next) {
+    next.name !== this.props.name &&
+    setTimeout(() => this.model.broadcast())
+  }
+  render(props) {
+    if (_.isFunction(this.props.render) ||
+        _.isFunction(this.props.children) ||
+        _.isFunction(this.props.component)) {
+      return super.render(props || { fields: this.model })
+    }
+    return super.render(props || {})
+  }
+  static get displayName() {
+    return 'FieldSet'
+  }
+  static get childContextTypes() {
+    return this.contextTypes
   }
   static get propTypes() {
     return {
       ...super.propTypes,
-      init: object,
-      component: oneOfType([string, func]).isRequired
+      init: PropTypes.object.isRequired,
+      children: PropTypes.oneOfType([
+        PropTypes.array, PropTypes.func, PropTypes.element
+      ])
     }
   }
   static get defaultProps() {
@@ -42,61 +211,3 @@ export class FieldSet extends SuperControl.View {
     }
   }
 }
-
-export class FieldSetModel extends SuperControl.Model {
-  constructor(...args) {
-    super(...args)
-    this.fields = {}
-    this.checkAll = this.checkAll.bind(this)
-  }
-  get visited() {
-    return this.form.getVisited(this.path, {})
-  }
-  get touched() {
-    return this.form.getTouched(this.path, {})
-  }
-  get isTouched() {
-    return _.someValues(this.touched, _.id)
-  }
-  toState() {
-    return _.pick(this, [
-      'init', 'value', 'touched', 'error', 'notice', 'visited'
-    ])
-  }
-  toProp() {
-    const name = this.path.pop()
-    const state = this.toState()
-    const { form, isTouched } = this
-    const isValid = !state.error
-    const isInvalid = !isValid
-    const isDirty = !_.deepEqual(state.init, state.value)
-    const isPristine = !isDirty
-    return _.assign(state, {
-      form, name, isValid, isInvalid, isDirty, isPristine, isTouched
-    })
-  }
-  register(field, [ key, ...path ]) {
-    this.fields = path.length
-      ? _.set(this.fields, [key], this.fields[key].register(field, path))
-      : _.set(this.fields, [key], field)
-    return this
-  }
-  check(value, values, method, [ key, ...path ]) {
-    return _.assign(
-      super.check(_.set(this.value, [key, ...path], value), values, method),
-      this.fields[key].check(value, values, method, path)
-    )
-  }
-  checkAll(value, values, method) {
-    return _.keys(this.fields)
-      .reduce((checked, key) => {
-        const { check, checkAll } = this.fields[key]
-        return _.assign(
-          checked,
-          (checkAll || check)(values[key], values, method)
-        )
-      }, { [this.id]: this[`_${method}`](value, values) || null })
-  }
-}
-
-export const modelFieldSet = (...args) => new FieldSetModel(...args)
