+++ src/super-control.js
@@ -1,99 +1,226 @@
-import { Component } from 'react'
-import { string, shape, func, number, oneOfType } from 'prop-types'
+import { Component, createElement } from 'react'
+import PropTypes from 'prop-types'
 import * as _ from './util'
 
-export class View extends Component {
-  constructor(...args) {
-    super(...args)
-    this.register = this.register.bind(this)
+export const Model = class SuperControlModel {
+  constructor(form, init, route, config) {
+    const _state = {
+      visits: 0,
+      touches: 0,
+      init: init,
+      value: init,
+      error: null,
+      notice: null,
+      isActive: false
+    }
+    _.assign(this, {
+      form,
+      route,
+      _state,
+      config,
+      subscribers: [],
+      publish: this.publish.bind(this),
+      validate: this.validate.bind(this),
+      initialize: this.initialize.bind(this)
+    })
+  }
+  get name() {
+    return (this.route.length || null) &&
+           this.route[this.route.length - 1]()
+  }
+  get names() {
+    return this.route.map(_.invoke)
+  }
+  get path() {
+    return _.toPath(this.names)
+  }
+  get init() {
+    return this._state.init
+  }
+  get value() {
+    return this._state.value
+  }
+  get error() {
+    return this._state.error || null
+  }
+  get notice() {
+    return this._state.notice || null
+  }
+  get hasNotice() {
+    return !!this._state.notice
+  }
+  get isInvalid() {
+    return !!this._state.error
+  }
+  get isActive() {
+    return this._state.isActive
+  }
+  get isVisited() {
+    return !!this._state.visits
+  }
+  get isTouched() {
+    return !!this._state.touches
+  }
+  getState() {
+    return _.pick(this, [
+      'name', 'path', 'init', 'value',
+      'isVisited', 'isTouched', 'isActive',
+      'error', 'notice', 'isInvalid', 'hasNotice'
+    ])
+  }
+  subscribe(subscriber) {
+    const index = this.subscribers.push(subscriber) - 1
+    subscriber(this.getState())
+    return _ => {
+      this.subscribers.splice(index, 1)
+      !this.subscribers.length && this.form.unregister(this.names, this)
+    }
+  }
+  publish() {
+    const state = this.getState()
+    this.subscribers.forEach(subscriber => subscriber(state))
+    return this
+  }
+  _setState(next, { silent = false } = {}) {
+    this._state = next
+    return silent ? this : this.publish()
+  }
+  _patch(change, options = {}) {
+    const { form, config, _state: current } = this
+    const next = _.assign({}, current, _.omit(change, ['visits', 'touches']))
+    if (change.visits || change.touches) {
+      if (change.visits) next.visits += change.visits
+      if (change.touches) next.touches += change.touches
+      next.isActive = (next.visits > current.visits) ||
+                      (current.isActive && next.touches <= current.touches)
+    }
+    if (options.notify) {
+      next.notice = config.notify(next.value, form.values, this)
+    }
+    if (options.validate) {
+      next.error = config.validate(next.value, form.values, this)
+    }
+    return this._setState(next, options)
+  }
+  initialize(init) {
+    this.form.isInitialized = false
+    this.form._patch(this.names, { init, value: init }, { quiet: true })
+    this.form.isInitialized = true
+  }
+  validate(options) {
+    const { form, config, value } = this
+    this._patch({ error: config.validate(value, form.values, this) }, options)
+    return this
+  }
+  static create(form, init = null, route = [], config = {}) {
+    return new this(form, init, route, _.defaults({}, config, {
+      override: _.id,
+      notify: _.toNull,
+      validate: _.toNull
+    }))
+  }
+}
+
+export const View = class SuperControlView extends Component {
+  get init() {
+    return this.props.init
+  }
+  get Model() {
+    return Model
+  }
+  get config() {
+    return _.pick(this.props, ['notify', 'validate'])
+  }
+  equalProps(current, next) {
+    const ignored = ['name', 'init']
+    _.isFunction(current.children) &&
+       _.isFunction(next.children) &&
+          ignored.push('children')
+    return _.shallowEqual(_.omit(current, ignored), _.omit(next, ignored))
+  }
+  equalState(current, next) {
+    return _.shallowEqual(current, next)
   }
   componentWillMount() {
     this.model = this.context['@@super-controls'].register({
-      init: this.getInit(),
-      paths: [_ => this.props.name],
-      model: (...args) => this.modelField(...args, {
-        notify: this.props.notify,
-        validate: this.props.validate
-      })
+      route: [_ => this.props.name],
+      ..._.pick(this, ['init', 'Model', 'config'])
     })
-    this.setState(this.model.toState())
+    this.unsubscribe = this.model.subscribe(this.setState.bind(this))
+  }
+  componentWillReceiveProps(next) {
+    next.name !== this.props.name &&
+    setTimeout(() => this.model.publish())
   }
   shouldComponentUpdate(nextProps, nextState) {
-    return !_.equalProps(this.props, nextProps) ||
-           !_.deepEqual(this.model.toState(), nextState)
-  }
-  componentDidUpdate() {
-    this.setState(this.model.toState())
+    return !this.equalProps(this.props, nextProps) ||
+           !this.equalState(this.state, nextState)
   }
   componentWillUnmount() {
-    this.model.unregister()
+    this.unsubscribe = this.unsubscribe()
   }
-  register({ init, model, paths }) {
-    return this.context['@@super-controls'].register({
-      init,
-      model,
-      paths: [_ => this.props.name, ...paths]
+  render(props = {}) {
+    const { render, component, children } = this.props
+    if (_.isFunction(render)) {
+      return render({
+        ..._.omit(this.props, ['render']),
+        ...props
+      })
+    }
+    if (_.isFunction(children)) {
+      return children({
+        ..._.omit(this.props, ['children']),
+        ...props
+      })
+    }
+    if (_.isString(component)) {
+      return createElement(component, {
+        ..._.omit(this.props, [
+          'init', 'render', 'component',
+          'parse', 'format', 'override',
+          'notify', 'validate'
+        ]),
+        ...props
+      })
+    }
+    return createElement(component, {
+      ..._.omit(this.props, ['component']),
+      ...props
     })
   }
-  getChildContext() {
-    return { '@@super-controls': { register: this.register } }
+  static get displayName() {
+    return 'SuperControl'
+  }
+  static get contextTypes() {
+    return {
+      '@@super-controls': PropTypes.shape({
+        register: PropTypes.func.isRequired
+      })
+    }
   }
   static get propTypes() {
     return {
-      notify: func,
-      validate: func,
-      name: oneOfType([string, number]).isRequired
+      init: PropTypes.any,
+      render: PropTypes.func,
+      notify: PropTypes.func,
+      validate: PropTypes.func,
+      name: PropTypes.oneOfType([
+        PropTypes.string, PropTypes.number
+      ]).isRequired,
+      children: PropTypes.oneOfType([
+        PropTypes.array, PropTypes.func
+      ]),
+      component: PropTypes.oneOfType([
+        PropTypes.string, PropTypes.func
+      ])
     }
   }
   static get defaultProps() {
     return {
+      init: null,
       notify: _.noop,
       validate: _.noop,
-      component: _ => null
+      component: _.toNull
     }
-  }
-  static get contextTypes() {
-    return {
-      '@@super-controls': shape({
-        register: func.isRequired
-      })
-    }
-  }
-  static get childContextTypes() {
-    return this.contextTypes
-  }
-}
-
-export class Model {
-  constructor(id, form, init, paths, { notify, validate, override } = {}) {
-    this.id = id
-    this.form = form
-    this._init = init
-    this._path = paths
-    this._notify = notify || _.noop
-    this._validate = validate || _.noop
-    this._override = override || _.id
-    this.check = this.check.bind(this)
-  }
-  get path() {
-    return this._path.map(_.invoke)
-  }
-  get init() {
-    return this.form.getInit(this.path, this._init)
-  }
-  get value() {
-    return this.form.getValue(this.path, this.init)
-  }
-  get error() {
-    return this.form.getError(this.id)
-  }
-  get notice() {
-    return this.form.getNotice(this.id)
-  }
-  check(value, values, method) {
-    return { [this.id]: this[`_${method}`](value, values) || null }
-  }
-  unregister() {
-    this.form.unregister(this)
   }
 }
